+++ test/common/audit.test.js
@@ -19,13 +19,14 @@ import nock from 'nock';
 import { createSite } from '@adobe/spacecat-shared-data-access/src/models/site.js';
 import { createOrganization } from '@adobe/spacecat-shared-data-access/src/models/organization.js';
 import { createConfiguration } from '@adobe/spacecat-shared-data-access/src/models/configuration.js';
-import { composeAuditURL, prependSchema } from '@adobe/spacecat-shared-utils';
+import { composeAuditURL, hasText, prependSchema } from '@adobe/spacecat-shared-utils';
 import {
   defaultMessageSender,
   defaultOrgProvider,
   defaultPersister,
   defaultSiteProvider,
   defaultUrlResolver,
+  noopPersister,
   noopUrlResolver,
   wwwUrlResolver,
 } from '../../src/common/audit.js';
@@ -135,8 +136,7 @@ describe('Audit tests', () => {
 
       await defaultMessageSender(resultMessage, context);
 
-      expect(context.sqs.sendMessage).to.have.been.calledOnce;
-      expect(context.sqs.sendMessage).to.have.been.calledWith(queueUrl, resultMessage);
+      expect(context.sqs.sendMessage).not.to.have.been.calledOnce;
     });
 
     it('default url resolves gets the base url and follows redirects', async () => {
@@ -268,14 +268,7 @@ describe('Audit tests', () => {
       expect(context.dataAccess.addAudit).to.have.been.calledWith(auditData);
 
       const finalUrl = 'space.cat';
-      const expectedMessage = {
-        type: message.type,
-        url: 'https://space.cat',
-        auditContext: { someField: 431, finalUrl, fullAuditRef },
-        auditResult: { metric: 42 },
-      };
-      expect(context.sqs.sendMessage).to.have.been.calledOnce;
-      expect(context.sqs.sendMessage).to.have.been.calledWith(queueUrl, expectedMessage);
+      expect(context.sqs.sendMessage).not.to.have.been.calledOnce;
       expect(postProcessors[0]).to.have.been.calledWith(finalUrl, auditData);
       expect(postProcessors[1]).to.have.been.calledWith(finalUrl, auditData);
       expect(postProcessors[2]).to.not.have.been.called;
@@ -332,14 +325,7 @@ describe('Audit tests', () => {
       id: 'some-audit-id',
     });
 
-    const expectedMessage = {
-      type: message.type,
-      url: 'https://space.cat',
-      auditContext: { finalUrl: 'space.cat', fullAuditRef },
-      auditResult: { metric: 42 },
-    };
-    expect(context.sqs.sendMessage).to.have.been.calledOnce;
-    expect(context.sqs.sendMessage).to.have.been.calledWith(queueUrl, expectedMessage);
+    expect(context.sqs.sendMessage).not.to.have.been.calledOnce;
   });
 
   it('wwwUrlResolver calculates audit urls correctly', async () => {
@@ -349,5 +335,17 @@ describe('Audit tests', () => {
     expect(wwwUrlResolver(createSite({ baseURL: 'https://www.spacecat.com' }))).to.equal('www.spacecat.com');
     expect(wwwUrlResolver(createSite({ baseURL: 'http://blog.spacecat.com' }))).to.equal('blog.spacecat.com');
     expect(wwwUrlResolver(createSite({ baseURL: 'https://blog.spacecat.com' }))).to.equal('blog.spacecat.com');
+  });
+
+  it('noop persister', async () => {
+    const audit = await noopPersister({
+      siteId: 'site-id',
+      isLive: true,
+      auditedAt: new Date().toISOString(),
+      auditType: 'some-type',
+      auditResult: { metric: 42 },
+      fullAuditRef: 'hmm',
+    });
+    expect(hasText(audit.getId())).to.be.true;
   });
 });